+++ service-datamodel/src/main/java/fi/nls/oskari/fe/gml/util/GeometryProperty.java
@@ -1,5 +1,7 @@
 package fi.nls.oskari.fe.gml.util;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.vividsolutions.jts.geom.Geometry;
 
 import fi.nls.oskari.fe.xml.util.Reference;
@@ -8,11 +10,9 @@ import fi.nls.oskari.fe.xml.util.Reference;
  * NOTE: CustomDeserializer means any attributes etc are lost by default
  */
 public class GeometryProperty extends Reference {
-
     private Geometry geometry;
 
     public GeometryProperty() {
-
     }
 
     public GeometryProperty(Geometry value) {
@@ -20,14 +20,12 @@ public class GeometryProperty extends Reference {
 
     }
 
-    @org.codehaus.jackson.annotate.JsonIgnore
-    @com.fasterxml.jackson.annotation.JsonIgnore
+    @JsonIgnore
     public Geometry getGeometry() {
         return geometry;
     }
 
-    @com.fasterxml.jackson.annotation.JsonProperty("WKT")
-    @org.codehaus.jackson.annotate.JsonProperty("WKT")
+    @JsonProperty("WKT")
     public String getWKT() {
         return geometry != null ? geometry.toText() : "";
     }