+++ service-feature-engine/src/main/java/fi/nls/oskari/fe/input/jackson/GeometryPropertyDeserializer.java
@@ -43,7 +43,11 @@ public class GeometryPropertyDeserializer extends
                 gml));
         return handlers;
     }
-
+/*
+    public Class<?> handledType() {
+        return GeometryProperty.class;
+    }
+*/
     public Map<QName, PullParserHandler> getHandlers() {
 
             return handlers;
@@ -56,9 +60,6 @@ public class GeometryPropertyDeserializer extends
             handlers.put(qname, new FEPullParser.ElementPullParserHandler(
                     qname, gml));
         }
-
-        ;
-
     }
 
 
@@ -101,8 +102,7 @@ public class GeometryPropertyDeserializer extends
 
     @Override
     public GeometryProperty deserialize(JsonParser jp,
-            DeserializationContext ctxt) throws IOException,
-            JsonProcessingException {
+            DeserializationContext ctxt) throws IOException {
         Geometry geom = null;
 
         FromXmlParser parser = (FromXmlParser) ctxt.getParser();