+++ service-feature-engine/src/main/java/fi/nls/oskari/fe/input/jackson/GmlMapper.java
@@ -1,11 +1,9 @@
 package fi.nls.oskari.fe.input.jackson;
 
 import java.util.Calendar;
-import java.util.Map;
 
-import javax.xml.namespace.QName;
-
-import org.apache.commons.lang3.tuple.Pair;
+import com.fasterxml.jackson.core.Version;
+import com.fasterxml.jackson.databind.module.SimpleModule;
 import org.geotools.xml.Configuration;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
@@ -13,15 +11,12 @@ import com.fasterxml.jackson.dataformat.xml.JacksonXmlModule;
 
 import fi.nls.oskari.fe.gml.util.GeometryProperty;
 import fi.nls.oskari.fe.input.format.gml.FEPullParser;
-import fi.nls.oskari.fe.input.format.gml.FEPullParser.PullParserHandler;
-import fi.nls.oskari.fe.iri.Resource;
 import fi.nls.oskari.fe.output.jackson.GeometryPropertySerializer;
-import fi.nls.oskari.fe.xml.util.NillableType;
 
 public class GmlMapper extends XmlMapper {
 
     /**
-     * 
+     *
      */
     private static final long serialVersionUID = -4030547605256325814L;
     private FEPullParser parserAny;
@@ -32,13 +27,13 @@ public class GmlMapper extends XmlMapper {
     static <T> Class getClazz(T... param) {
         return param.getClass().getComponentType();
     }
-    
+
     public GmlMapper(Configuration gml, boolean lenient) {
 
         super(new XmlFactory());
 
         JacksonXmlModule module = new JacksonXmlModule();
-        
+
         // ?
         module.setDefaultUseWrapper(false);
 
@@ -47,9 +42,15 @@ public class GmlMapper extends XmlMapper {
         geometryDeserializer = new GeometryPropertyDeserializer(gml, parserAny);
         calendarDeserializer = new CalendarDeserializer();
 
+        // Registering serializers to XmlModule doesn't seem to do the trick anymore on Jackson 2.x
+        final SimpleModule oskariGMLModule = new SimpleModule("OskariGMLModule", new Version(1, 0, 0, null))
+                .addDeserializer(GeometryProperty.class, geometryDeserializer)
+                .addDeserializer(Calendar.class, calendarDeserializer);
+        registerModule(oskariGMLModule);
+
         module.addDeserializer(GeometryProperty.class, geometryDeserializer);
         module.addDeserializer(Calendar.class,calendarDeserializer);
-        
+
         enable(DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY);
         if (lenient) {
             disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);