+++ servlet-transport/src/main/java/fi/nls/oskari/utils/GeometrySerializer.java
@@ -2,19 +2,16 @@ package fi.nls.oskari.utils;
 
 import java.io.IOException;
 
-import org.codehaus.jackson.JsonGenerator;
-import org.codehaus.jackson.JsonProcessingException;
-import org.codehaus.jackson.map.JsonSerializer;
-import org.codehaus.jackson.map.SerializerProvider;
-
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
 import com.vividsolutions.jts.geom.Geometry;
 
 class GeometrySerializer extends JsonSerializer<Geometry> {
 
     @Override
     public void serialize(Geometry value, JsonGenerator jgen,
-            SerializerProvider provider) throws IOException,
-            JsonProcessingException {
+            SerializerProvider provider) throws IOException {
         if (value == null) {
             provider.defaultSerializeNull(jgen);
             return;