+++ servlet-transport/src/test/java/fi/nls/oskari/work/fe/JacksonCounterJsonResultProcessor.java
@@ -3,25 +3,20 @@ package fi.nls.oskari.work.fe;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 
-import org.codehaus.jackson.JsonGenerationException;
-import org.codehaus.jackson.impl.DefaultPrettyPrinter;
-import org.codehaus.jackson.map.JsonMappingException;
-import org.codehaus.jackson.map.ObjectMapper;
-import org.codehaus.jackson.map.ObjectWriter;
-import org.codehaus.jackson.map.SerializationConfig;
-import org.codehaus.jackson.map.annotate.JsonSerialize.Inclusion;
-
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.core.JsonGenerationException;
+import com.fasterxml.jackson.core.util.DefaultPrettyPrinter;
+import com.fasterxml.jackson.databind.*;
+import fi.nls.oskari.fe.input.jackson.XmlMapper;
 import fi.nls.oskari.utils.GeometryJSONOutputModule;
 
 public class JacksonCounterJsonResultProcessor extends CounterJsonResultProcessor {
 
-    protected ObjectMapper json = new ObjectMapper();
+    protected ObjectMapper json = new XmlMapper();
     protected ObjectWriter writer;
 
     public JacksonCounterJsonResultProcessor() {
-        SerializationConfig x = json.getSerializationConfig()
-                .withSerializationInclusion(Inclusion.NON_NULL);
-        json.setSerializationConfig(x);
+        json.setSerializationInclusion(JsonInclude.Include.NON_NULL);
         writer = json.writer(new DefaultPrettyPrinter());
 
         GeometryJSONOutputModule simpleModule = new GeometryJSONOutputModule();