+++ ReleaseNotes.md
@@ -7,6 +7,12 @@
 Servlet-API upgraded from 2.4 to 3.1.0 in preparation of replacing current servlet-map/webapp-map with spring
  counterparts from oskari-spring repository.
 
+Other updates: 
+* Jackson 1.9.11 -> 2.5.4
+* Jedis 2.6.0 -> 2.7.2
+* org.codehaus.woodstox:stax2-api 3.1.1 -> 3.1.4
+* com.fasterxml.woodstox:woodstox-core 4.4.1 -> 5.0.1
+
 ### Property changes
 
 db.additional.pools has been changed to db.additional.modules to better describe it.